requirements.txt
lmdb==1.4.0
matplotlib==3.5.1
numpy==1.23.2
opencv_python==4.5.5.64
Pillow==9.4.0
PyWavelets==1.3.0
tensorboardX==2.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
wandb==0.13.11
